refactor: yml에 필요한 카프카 설정 추가
