+++ src/backend/push-server/build.gradle
@@ -54,6 +54,18 @@ dependencies {
 	// Protobuf
 	implementation 'com.google.protobuf:protobuf-java:3.21.12'
 
+	// Swagger
+	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.7.0'
+
+	// redis
+	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
+	// Kafka
+	implementation 'org.springframework.kafka:spring-kafka'
+
+	// JSON 처리
+	implementation 'com.fasterxml.jackson.core:jackson-databind'
+
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
