+++ src/backend/push-server/src/main/java/com/bbebig/pushserver/PushServerApplication.java
@@ -1,4 +1,4 @@
-package com.bbebig.push_server;
+package com.bbebig.pushserver;
 
 import io.github.cdimascio.dotenv.Dotenv;
 import org.springframework.boot.SpringApplication;
@@ -6,10 +6,15 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.ComponentScan;
 
 @EnableFeignClients
 @EnableDiscoveryClient
 @SpringBootApplication(exclude = {DataSourceAutoConfiguration.class})
+@ComponentScan(basePackages = {
+		"com.bbebig.pushserver",
+		"com.bbebig.commonmodule"
+})
 public class PushServerApplication {
 
 	public static void main(String[] args) {
