+++ src/backend/push-server/src/main/java/com/bbebig/pushserver/global/config/PushServerConfig.java
@@ -1,0 +1,12 @@
+package com.bbebig.pushserver.global.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = {
+		"com.bbebig.commonmodule.kafka.config",
+		"com.bbebig.commonmodule.redis.config"
+})
+public class PushServerConfig {
+}
